refactor: nextJS 스럽게 리팩토링
- 각 라우트 폴더 home/page 에는 export default 를 항상 가져야 하며 컴포넌트 이름은 page여야 한다.
- 따라서 home/page.tsx 내부에 function homeView로 선언한 컴포넌트들을 다 컴포넌트 폴더로 따로 이동시키고 한단계 더 function page 내에서 HomeView를 호출하도록 변경했다
